+++ bin/handler-statuspage.rb
@@ -13,10 +13,75 @@
 require 'sensu-handler'
 require 'redphone/statuspage'
 
+ALLOWED_COMPONENT_STATUSES = %w(operational degraded_performance partial_outage major_outage ignore).freeze
+DEFAULT_COMPONENT_STATUSES = {
+  'warning' => 'partial_outage',
+  'critical' => 'major_outage'
+}.freeze
+
+ALLOWED_INCIDENT_STATUSES = %w(investigating identified monitoring resolved ignore).freeze
+DEFAULT_INCIDENT_STATUSES = {
+  'warning' => 'identified',
+  'critical' => 'identified'
+}.freeze
+
 # main plugin class
 class StatusPage < Sensu::Handler
   def incident_key
     @event['client']['name'] + '/' + @event['check']['name']
+  end
+
+  def status_name
+    case @event['check']['status']
+    when 1
+      'warning'
+    when 2
+      'critical'
+    end
+  end
+
+  def component_status
+    case @event['action']
+    when 'create'
+      desired_status = @event['check']["statuspage_component_#{status_name}"]
+      if desired_status && ALLOWED_COMPONENT_STATUSES.include?(desired_status)
+        desired_status
+      else
+        DEFAULT_COMPONENT_STATUSES[status_name]
+      end
+    when 'resolve'
+      'operational'
+    end
+  end
+
+  def incident_status
+    case @event['action']
+    when 'create'
+      desired_status = @event['check']["statuspage_incident_#{status_name}"]
+      if desired_status && ALLOWED_INCIDENT_STATUSES.include?(desired_status)
+        desired_status
+      else
+        DEFAULT_INCIDENT_STATUSES[status_name]
+      end
+    when 'resolve'
+      'resolved'
+    end
+  end
+
+  def ignore_status?(status)
+    ['ignore', nil].include?(status)
+  end
+
+  def description
+    @event['notification'] || [@event['client']['name'], @event['check']['name'], @event['check']['output']].join(' : ')
+  end
+
+  def incident_message(status)
+    if status == 'resolved'
+      "Problem with #{incident_key} has been resolved."
+    else
+      "There has been a problem: #{description}."
+    end
   end
 
   def handle
@@ -24,47 +89,37 @@ class StatusPage < Sensu::Handler
       page_id: settings['statuspage']['page_id'],
       api_key: settings['statuspage']['api_key']
     )
-    description = @event['notification'] || [@event['client']['name'], @event['check']['name'], @event['check']['output']].join(' : ')
     begin
       timeout(3) do
         if @event['check'].key?('component_id')
-          status = case @event['action']
-                   when 'create'
-                     'major_outage'
-                   when 'resolve'
-                     'operational'
-                   end
-          unless status.nil?
+          unless ignore_status?(component_status)
             statuspage.update_component(
               component_id: @event['check']['component_id'],
-              status: status
+              status: component_status
             )
           end
         end
-        response = case @event['action']
-                   when 'create'
+        existing_incident = statuspage.get_all_incidents.detect { |incident| incident['name'] == incident_key && incident['status'] != 'resolved' }
+        response = if existing_incident
+                     statuspage.update_incident(
+                       name: incident_key,
+                       wants_twitter_update: 'f',
+                       status: incident_status,
+                       incident_id: existing_incident['id'],
+                       message: incident_message(incident_status)
+                     )
+                   elsif !ignore_status?(incident_status)
                      statuspage.create_realtime_incident(
                        name: incident_key,
-                       status: 'investigating',
+                       status: incident_status,
                        wants_twitter_update: 'f',
-                       message: "There has been a problem: #{description}."
-                     )
-                   when 'resolve'
-                     incident_id = nil
-                     statuspage.get_all_incidents.each do |incident|
-                       if incident['name'] == incident_key
-                         incident_id = incident['id']
-                         break
-                       end
-                     end
-                     statuspage.update_incident(
-                       name: "Problem with #{incident_key} has been resolved.",
-                       wants_twitter_update: 'f',
-                       status: 'resolved',
-                       incident_id: incident_id
+                       message: incident_message(incident_status)
                      )
                    end
-        if (response['status'] == 'investigating' && @event['action'] == 'create') || (response['status'] == 'resolved' && @event['action'] == 'resolve')
+        if ignore_status?(incident_status) ||
+           (response['status'] == incident_status && @event['action'] == 'create') ||
+           (response['status'] == 'resolved'      && @event['action'] == 'resolve')
+
           puts 'statuspage -- ' + @event['action'].capitalize + 'd incident -- ' + incident_key
         else
           puts 'statuspage -- failed to ' + @event['action'] + ' incident -- ' + incident_key